+++ code/modules/modular_computers/computers/item/role_tablet_presets.dm
@@ -326,6 +326,9 @@
 	icon_state = "/obj/item/modular_computer/pda/botanist"
 	greyscale_config = /datum/greyscale_config/tablet/stripe_thick
 	greyscale_colors = "#50E193#E26F41#71A7CA"
+	starting_programs = list(
+		/datum/computer_file/program/botanical_encyclopedia
+	)
 
 /obj/item/modular_computer/pda/cook
 	name = "cook PDA"
