+++ tgstation.dme
@@ -573,6 +573,7 @@
 #include "code\_globalvars\lists\achievements.dm"
 #include "code\_globalvars\lists\ambience.dm"
 #include "code\_globalvars\lists\basic_ai.dm"
+#include "code\_globalvars\lists\botany.dm"
 #include "code\_globalvars\lists\canisters.dm"
 #include "code\_globalvars\lists\cargo.dm"
 #include "code\_globalvars\lists\client.dm"
@@ -3740,7 +3741,6 @@
 #include "code\modules\asset_cache\assets\research_designs.dm"
 #include "code\modules\asset_cache\assets\rtd.dm"
 #include "code\modules\asset_cache\assets\safe.dm"
-#include "code\modules\asset_cache\assets\seeds.dm"
 #include "code\modules\asset_cache\assets\sheetmaterials.dm"
 #include "code\modules\asset_cache\assets\slot_machines.dm"
 #include "code\modules\asset_cache\assets\supplypods.dm"
@@ -5850,6 +5850,7 @@
 #include "code\modules\modular_computers\file_system\programs\atmosscan.dm"
 #include "code\modules\modular_computers\file_system\programs\betting.dm"
 #include "code\modules\modular_computers\file_system\programs\borg_monitor.dm"
+#include "code\modules\modular_computers\file_system\programs\botanical_encyclopedia.dm"
 #include "code\modules\modular_computers\file_system\programs\bounty_board.dm"
 #include "code\modules\modular_computers\file_system\programs\budgetordering.dm"
 #include "code\modules\modular_computers\file_system\programs\card.dm"
